Arreglar `make clippy`, que master vuelve a tener roto
Dos errores, ninguno mío, que entran en la rama al traer master:

- `e1000e.rs`: `as u32` sobre una expresión que ya es `u32`. Este cae
  en `zcore-drivers`, que es dependencia de todo lo demás, así que
  tapaba los avisos de los crates de arriba: `cargo clippy -p
  linux-object` moría aquí y ni siquiera llegaba a mirar
  `linux-object`.
- `syncobj_file.rs`: un `if ... .is_none() { return None }` que clippy
  quiere como `?`. Es el que rompe de verdad `make clippy ARCH=x86_64`,
  y con él el job `test-bare-metal`.

Con los dos, `make clippy ARCH=x86_64` y `cargo clippy -p zcore
--features "linux libos"` vuelven a pasar.

+++ drivers/src/net/e1000e.rs
@@ -1468,7 +1468,7 @@ impl E1000eHw {
     /// Bank detect via signature word 0x13 (Linux `e1000_valid_nvm_bank_detect_ich8lan` SPT).
     unsafe fn ich_flash_active_bank_words(&self) -> u32 {
         let strap = mmio_read(self.base, E1000E_STRAP);
-        let nvm_size = ((((strap >> 1) & 0x1F) + 1) * 4096) as u32;
+        let nvm_size = (((strap >> 1) & 0x1F) + 1) * 4096;
         let bank_words = (nvm_size / 2) / 2; // words per bank
 
         // Signature at word offset 0x13 of each bank; valid if (sig & 0xC0) == 0x80.
